chore: added requested changes
