+++ bc_obps/reporting/service/report_supplementary_version_service/__init__.py
@@ -1,4 +1,3 @@
 # ruff: noqa: F401
 
 from .report_supplementary_version_service import ReportSupplementaryVersionService
-from .report_supplementary_cloning import ReportSupplementaryCloning
