+++ bc_obps/reporting/tests/service/test_report_supplementary_service/test_report_supplementary_cloning.py
@@ -6,11 +6,17 @@ from django.utils import timezone
 from registration.models import Operation
 from reporting.models.report_raw_activity_data import ReportRawActivityData
 from reporting.tests.service.test_report_activity_save_service import data
-from reporting.service.report_supplementary_version_service.report_supplementary_version_service import (
-    ReportSupplementaryVersionService,
-)
 from reporting.service.report_supplementary_version_service.report_supplementary_cloning import (
-    ReportSupplementaryCloning,
+    clone_report_version_operation,
+    clone_report_version_representatives,
+    clone_report_version_person_responsible,
+    clone_report_version_additional_data,
+    clone_report_version_new_entrant_data,
+    clone_report_version_verification,
+    clone_report_version_attachments,
+    clone_report_version_facilities,
+    reapply_emission_categories,
+    clone_all,
 )
 from reporting.models import (
     FacilityReport,
@@ -32,87 +38,8 @@ from reporting.models import (
 import common.lib.pgtrigger as pgtrigger
 
 
-class ReportSupplementaryVersionServiceTests(TestCase):
+class ReportSupplementaryCloningTests(TestCase):
     def setUp(self):
-        # Past and Current reporting year setup
-        self.operator = make_recipe("registration.tests.utils.operator")
-        self.operation = make_recipe(
-            "registration.tests.utils.operation",
-            operator=self.operator,
-            registration_purpose=Operation.Purposes.OBPS_REGULATED_OPERATION,
-        )
-
-        self.current_reporting_year = make_recipe('reporting.tests.utils.reporting_year', reporting_year=2030)
-        self.past_reporting_year = make_recipe('reporting.tests.utils.reporting_year', reporting_year=2029)
-
-        self.past_year_old_report = make_recipe(
-            'reporting.tests.utils.report', operator=self.operator, operation=self.operation, reporting_year_id=2029
-        )
-        self.past_year_old_report_version = make_recipe(
-            'reporting.tests.utils.report_version',
-            status=ReportVersion.ReportVersionStatus.Draft,
-            report=self.past_year_old_report,
-        )
-
-        self.past_year_old_report_operation = make_recipe(
-            "reporting.tests.utils.report_operation",
-            report_version=self.past_year_old_report_version,
-            operator_legal_name=self.operator.legal_name,
-            registration_purpose=Operation.Purposes.OBPS_REGULATED_OPERATION,
-        )
-
-        self.past_year_representative = make_recipe(
-            'reporting.tests.utils.report_operation_representative',
-            report_version=self.past_year_old_report_version,
-            representative_name="Current Rep",
-        )
-        self.past_year_person_responsible = make_recipe(
-            'reporting.tests.utils.report_person_responsible',
-            report_version=self.past_year_old_report_version,
-        )
-        self.past_year_additional_data = make_recipe(
-            'reporting.tests.utils.report_additional_data',
-            report_version=self.past_year_old_report_version,
-            capture_emissions=True,
-            electricity_generated=999.0,
-        )
-        self.past_year_facility_report = make_recipe(
-            'reporting.tests.utils.facility_report',
-            report_version=self.past_year_old_report_version,
-        )
-
-        self.current_year_report = make_recipe(
-            'reporting.tests.utils.report', operator=self.operator, operation=self.operation, reporting_year_id=2030
-        )
-        self.current_year_report_version = make_recipe(
-            'reporting.tests.utils.report_version',
-            status=ReportVersion.ReportVersionStatus.Draft,
-            report=self.current_year_report,
-        )
-        self.current_year_report_operation = make_recipe(
-            "reporting.tests.utils.report_operation", report_version=self.current_year_report_version
-        )
-
-        self.current_year_representative = make_recipe(
-            'reporting.tests.utils.report_operation_representative',
-            report_version=self.current_year_report_version,
-            representative_name="Current Rep",
-        )
-        self.current_year_person_responsible = make_recipe(
-            'reporting.tests.utils.report_person_responsible',
-            report_version=self.current_year_report_version,
-        )
-        self.current_year_additional_data = make_recipe(
-            'reporting.tests.utils.report_additional_data',
-            report_version=self.current_year_report_version,
-            capture_emissions=True,
-            electricity_generated=999.0,
-        )
-        self.current_year_facility_report = make_recipe(
-            'reporting.tests.utils.facility_report',
-            report_version=self.current_year_report_version,
-        )
-
         # Create old and new ReportVersion instances
         self.old_report_version = make_recipe(
             'reporting.tests.utils.report_version',
@@ -230,212 +157,28 @@ class ReportSupplementaryVersionServiceTests(TestCase):
             status=ReportVersion.ReportVersionStatus.Draft,
         )
 
-    def test_create_report_supplementary_version(self):
-        # ACT: Call the method to create a supplementary version.
-        self.new_report_version.status = ReportVersion.ReportVersionStatus.Submitted
-        self.new_report_version.save()
-        new_version = ReportSupplementaryVersionService._create_supplementary_version(self.old_report_version)
-
-        # ASSERT: Verify that the new report version is correctly created.
-        self.assertEqual(
-            new_version.report,
-            self.old_report_version.report,
-            "The new report version should be associated with the same report as the original.",
-        )
-        self.assertEqual(
-            new_version.report_type,
-            self.old_report_version.report_type,
-            "The new report version should have the same report type as the original.",
-        )
-        self.assertEqual(
-            new_version.status,
-            ReportVersion.ReportVersionStatus.Draft,
-            "The new report version should be created with status Draft.",
-        )
-        self.assertFalse(
-            new_version.is_latest_submitted,
-            "The new report version should not be marked as the latest submitted version.",
-        )
-
-    def test_create_or_clone_same_purpose_no_operator_change(self):
-        """Test that when purpose hasn't changed and operator hasn't changed, supplementary version is created."""
-        operation = self.old_report_version.report.operation
-        operation.registration_purpose = Operation.Purposes.OBPS_REGULATED_OPERATION
-        operation.save()
-
-        self.old_report_version.report.operator = operation.operator
-        self.old_report_version.report.save()
-
-        self.new_report_version.status = ReportVersion.ReportVersionStatus.Submitted
-        self.new_report_version.save()
-
-        new_version = ReportSupplementaryVersionService.create_or_clone_report_version(self.old_report_version.id)
-
-        self.assertEqual(new_version.report, self.old_report_version.report)
-        self.assertIsNotNone(ReportOperation.objects.filter(report_version=new_version).first())
-
-    @patch('service.reporting_year_service.ReportingYearService.get_current_reporting_year')
-    def test_create_or_clone_purpose_changed_past_year_creates_supplementary(self, mock_get_year):
-        """Test that when registration purpose has changed since the past report was submitted,
-        a supplementary (cloned) version is created for the past year."""
-        self.past_year_old_report_version.status = ReportVersion.ReportVersionStatus.Submitted
-        self.past_year_old_report_version.save()
-
-        self.operation.registration_purpose = Operation.Purposes.OPTED_IN_OPERATION
-        self.operation.save()
-
-        mock_get_year.return_value = self.current_reporting_year
-
-        new_past_year_report_version = ReportSupplementaryVersionService.create_or_clone_report_version(
-            self.past_year_old_report_version.id
-        )
-
-        self.assertEqual(
-            new_past_year_report_version.report_operation.registration_purpose,
-            Operation.Purposes.OBPS_REGULATED_OPERATION,
-        )
-
-        # ReportOperation is cloned from the submitted version (old purpose/operator details)
-        new_report_op = new_past_year_report_version.report_operation
-        self.assertEqual(new_report_op.registration_purpose, Operation.Purposes.OBPS_REGULATED_OPERATION)
-        self.assertEqual(new_report_op.operator_legal_name, self.operator.legal_name)
-
-        # ReportOperationRepresentative is cloned
-        new_reps = ReportOperationRepresentative.objects.filter(report_version=new_past_year_report_version)
-        self.assertEqual(new_reps.count(), 1)
-        self.assertEqual(new_reps.first().representative_name, self.past_year_representative.representative_name)
-
-        # ReportPersonResponsible is cloned
-        new_person = ReportPersonResponsible.objects.filter(report_version=new_past_year_report_version).first()
-        self.assertIsNotNone(new_person, "Cloned version should have a ReportPersonResponsible.")
-        self.assertEqual(new_person.first_name, self.past_year_person_responsible.first_name)
-        self.assertEqual(new_person.last_name, self.past_year_person_responsible.last_name)
-
-        # ReportAdditionalData is cloned
-        new_additional = ReportAdditionalData.objects.filter(report_version=new_past_year_report_version).first()
-        self.assertIsNotNone(new_additional, "Cloned version should have ReportAdditionalData.")
-        self.assertEqual(new_additional.capture_emissions, self.past_year_additional_data.capture_emissions)
-        self.assertEqual(new_additional.electricity_generated, self.past_year_additional_data.electricity_generated)
-
-        # FacilityReport is cloned
-        new_facility = FacilityReport.objects.filter(report_version=new_past_year_report_version).first()
-        self.assertIsNotNone(new_facility, "Cloned version should have a FacilityReport.")
-        self.assertEqual(new_facility.facility_id, self.past_year_facility_report.facility_id)
-
-    @patch('service.report_version_service.ReportVersionService.create_report_version')
-    @patch('service.reporting_year_service.ReportingYearService.get_current_reporting_year')
-    def test_create_or_clone_purpose_changed_current_year_creates_blank(self, mock_get_year, mock_create):
-        """Test that for current years with purpose change, a blank version is created (no cloned data)."""
-
-        # Mock the current reporting year to match the current year report's year
-        mock_get_year.return_value = self.current_reporting_year
-
-        self.current_year_report_version.status = ReportVersion.ReportVersionStatus.Submitted
-        self.current_year_report_version.save()
-
-        # Change the operation's registration purpose so purpose_changed=True
-        self.operation.registration_purpose = Operation.Purposes.NEW_ENTRANT_OPERATION
-        self.operation.save()
-
-        new_current_year_report_version = ReportSupplementaryVersionService.create_or_clone_report_version(
-            self.current_year_report_version.id
-        )
-
-        self.assertEqual(
-            new_current_year_report_version.report_operation.registration_purpose,
-            Operation.Purposes.NEW_ENTRANT_OPERATION,
-        )
-
-        self.assertIsNone(
-            ReportPersonResponsible.objects.filter(report_version=new_current_year_report_version).first(),
-            "Blank version should have no ReportPersonResponsible.",
-        )
-        self.assertIsNone(
-            ReportAdditionalData.objects.filter(report_version=new_current_year_report_version).first(),
-            "Blank version should have no ReportAdditionalData.",
-        )
-        self.assertFalse(
-            FacilityReport.objects.filter(report_version=new_current_year_report_version).exists(),
-            "Blank version should have no FacilityReport.",
-        )
-        self.assertFalse(
-            ReportOperationRepresentative.objects.filter(report_version=new_current_year_report_version).exists(),
-            "Blank version should have no ReportOperationRepresentative.",
-        )
-
-    def test_create_or_clone_operator_changed_purpose_changed_creates_supplementary_with_same_details(self):
-        """Test that when operator and purpose both change, supplementary version is created cloning the submitted version's details."""
-        # --- Arrange: populate current_year_report_version with data to be cloned ---
-
-        # Update the existing bare report_operation with meaningful values
-        self.current_year_report_operation.operator_legal_name = self.operator.legal_name
-        self.current_year_report_operation.operator_trade_name = self.operator.trade_name
-        self.current_year_report_operation.operation_name = self.operation.name
-        self.current_year_report_operation.operation_type = Operation.Types.SFO
-        self.current_year_report_operation.registration_purpose = Operation.Purposes.OBPS_REGULATED_OPERATION
-        self.current_year_report_operation.save()
-
-        self.current_year_report_version.status = ReportVersion.ReportVersionStatus.Submitted
-        self.current_year_report_version.is_latest_submitted = True
-        self.current_year_report_version.save()
-
-        # --- Act: change operator and purpose, then create supplementary version ---
-        new_operator = make_recipe("registration.tests.utils.operator")
-        operation = self.current_year_report_version.report.operation
-        operation.registration_purpose = Operation.Purposes.NEW_ENTRANT_OPERATION
-        operation.operator = new_operator
-        operation.save()
-
-        new_current_year_report_version = ReportSupplementaryVersionService.create_or_clone_report_version(
-            self.current_year_report_version.id
-        )
-        # ReportOperation is cloned from the submitted version (old purpose/operator details)
-        new_report_op = new_current_year_report_version.report_operation
-        self.assertEqual(new_report_op.registration_purpose, Operation.Purposes.OBPS_REGULATED_OPERATION)
-        self.assertEqual(new_report_op.operator_legal_name, self.operator.legal_name)
-
-        # ReportOperationRepresentative is cloned
-        new_reps = ReportOperationRepresentative.objects.filter(report_version=new_current_year_report_version)
-        self.assertEqual(new_reps.count(), 1)
-        self.assertEqual(new_reps.first().representative_name, self.current_year_representative.representative_name)
-
-        # ReportPersonResponsible is cloned
-        new_person = ReportPersonResponsible.objects.filter(report_version=new_current_year_report_version).first()
-        self.assertIsNotNone(new_person, "Cloned version should have a ReportPersonResponsible.")
-        self.assertEqual(new_person.first_name, self.current_year_person_responsible.first_name)
-        self.assertEqual(new_person.last_name, self.current_year_person_responsible.last_name)
-
-        # ReportAdditionalData is cloned
-        new_additional = ReportAdditionalData.objects.filter(report_version=new_current_year_report_version).first()
-        self.assertIsNotNone(new_additional, "Cloned version should have ReportAdditionalData.")
-        self.assertEqual(new_additional.capture_emissions, self.current_year_additional_data.capture_emissions)
-        self.assertEqual(new_additional.electricity_generated, self.current_year_additional_data.electricity_generated)
-
-        # FacilityReport is cloned
-        new_facility = FacilityReport.objects.filter(report_version=new_current_year_report_version).first()
-        self.assertIsNotNone(new_facility, "Cloned version should have a FacilityReport.")
-        self.assertEqual(new_facility.facility_id, self.current_year_facility_report.facility_id)
-
     def test_clone_report_version_operation(self):
         """
         Test that the cloning method for ReportOperation correctly duplicates
         all scalar field values and many-to-many relationships (regulated_products and activities)
         from the old report version to a new report version.
         """
-        # PRE-ACT: Assert initial count of ReportOperation for old_report_version is 1.
-        initial_count = ReportOperation.objects.filter(report_version=self.old_report_version).count()
+        # PRE-ACT: Assert initial count of ReportOperation is 1.
+        initial_count = ReportOperation.objects.count()
         self.assertEqual(initial_count, 1, "There should be one ReportOperation initially.")
 
-        ReportSupplementaryCloning.clone_report_version_operation(self.old_report_version, self.new_report_version)
+        # ACT: Run the cloning method.
+        clone_report_version_operation(self.old_report_version, self.new_report_version)
 
-        new_count = ReportOperation.objects.filter(
-            report_version__in=[self.old_report_version, self.new_report_version]
-        ).count()
+        # ASSERT: Verify that a new ReportOperation has been created (count increases from 1 to 2).
+        new_count = ReportOperation.objects.count()
         self.assertEqual(new_count, 2, "A new ReportOperation should be created after cloning.")
 
+        # ASSERT: Retrieve both old and new ReportOperation instances.
         old_operation = ReportOperation.objects.get(report_version=self.old_report_version)
         new_operation = ReportOperation.objects.get(report_version=self.new_report_version)
 
+        # ASSERT: Check that scalar fields are copied correctly.
         self.assertEqual(new_operation.operator_legal_name, old_operation.operator_legal_name)
         self.assertEqual(new_operation.operator_trade_name, old_operation.operator_trade_name)
         self.assertEqual(new_operation.operation_name, old_operation.operation_name)
@@ -444,10 +187,12 @@ class ReportSupplementaryVersionServiceTests(TestCase):
         self.assertEqual(new_operation.operation_bcghgid, old_operation.operation_bcghgid)
         self.assertEqual(new_operation.bc_obps_regulated_operation_id, old_operation.bc_obps_regulated_operation_id)
 
+        # ASSERT: Verify many-to-many relationships for regulated_products are copied.
         old_regulated_ids = set(old_operation.regulated_products.values_list('id', flat=True))
         new_regulated_ids = set(new_operation.regulated_products.values_list('id', flat=True))
         self.assertEqual(old_regulated_ids, new_regulated_ids)
 
+        # ASSERT: Verify many-to-many relationships for activities are copied.
         old_activities_ids = set(old_operation.activities.values_list('id', flat=True))
         new_activities_ids = set(new_operation.activities.values_list('id', flat=True))
         self.assertEqual(old_activities_ids, new_activities_ids)
@@ -465,9 +210,8 @@ class ReportSupplementaryVersionServiceTests(TestCase):
             ReportOperationRepresentative.objects.filter(report_version=self.new_report_version).count(), 0
         )
 
-        ReportSupplementaryCloning.clone_report_version_representatives(
-            self.old_report_version, self.new_report_version
-        )
+        # ACT: Clone the representatives.
+        clone_report_version_representatives(self.old_report_version, self.new_report_version)
 
         # ASSERT: Verify that two new representatives have been created.
         new_reps = ReportOperationRepresentative.objects.filter(report_version=self.new_report_version)
@@ -491,9 +235,8 @@ class ReportSupplementaryVersionServiceTests(TestCase):
         self.assertIsNotNone(ReportPersonResponsible.objects.filter(report_version=self.old_report_version).first())
         self.assertIsNone(ReportPersonResponsible.objects.filter(report_version=self.new_report_version).first())
 
-        ReportSupplementaryCloning.clone_report_version_person_responsible(
-            self.old_report_version, self.new_report_version
-        )
+        # ACT: Clone the ReportPersonResponsible.
+        clone_report_version_person_responsible(self.old_report_version, self.new_report_version)
 
         # ASSERT: Retrieve and verify the cloned ReportPersonResponsible.
         new_person = ReportPersonResponsible.objects.filter(report_version=self.new_report_version).first()
@@ -518,9 +261,8 @@ class ReportSupplementaryVersionServiceTests(TestCase):
         self.assertIsNotNone(ReportAdditionalData.objects.filter(report_version=self.old_report_version).first())
         self.assertIsNone(ReportAdditionalData.objects.filter(report_version=self.new_report_version).first())
 
-        ReportSupplementaryCloning.clone_report_version_additional_data(
-            self.old_report_version, self.new_report_version
-        )
+        # ACT: Clone the ReportAdditionalData.
+        clone_report_version_additional_data(self.old_report_version, self.new_report_version)
 
         # ASSERT: Retrieve and verify the cloned additional data.
         new_additional_data = ReportAdditionalData.objects.filter(report_version=self.new_report_version).first()
@@ -554,9 +296,8 @@ class ReportSupplementaryVersionServiceTests(TestCase):
         self.assertGreater(old_productions_count, 0)
         self.assertIsNone(ReportNewEntrant.objects.filter(report_version=self.new_report_version).first())
 
-        ReportSupplementaryCloning.clone_report_version_new_entrant_data(
-            self.old_report_version, self.new_report_version
-        )
+        # ACT: Clone the ReportNewEntrant data.
+        clone_report_version_new_entrant_data(self.old_report_version, self.new_report_version)
 
         # ASSERT: Verify that a new ReportNewEntrant has been created (count increases from 0 to 1).
         new_new_entrant_instance = ReportNewEntrant.objects.filter(report_version=self.new_report_version).first()
@@ -616,7 +357,8 @@ class ReportSupplementaryVersionServiceTests(TestCase):
                 is_other_visit=False,
             )
 
-        ReportSupplementaryCloning.clone_report_version_verification(self.old_report_version, self.new_report_version)
+        # ACT: Clone the ReportVerification from old to new report version.
+        clone_report_version_verification(self.old_report_version, self.new_report_version)
 
         # ASSERT: Verify that a new ReportVerification has been created (record count increases).
         new_verification = ReportVerification.objects.filter(report_version=self.new_report_version).first()
@@ -645,10 +387,9 @@ class ReportSupplementaryVersionServiceTests(TestCase):
 
         ReportVerification.objects.filter(report_version=self.old_report_version).delete()
         # ACT: Attempt to clone ReportVerification.
-        ReportSupplementaryCloning.clone_report_version_verification(self.old_report_version, self.new_report_version)
+        clone_report_version_verification(self.old_report_version, self.new_report_version)
 
-        ReportSupplementaryCloning.clone_report_version_verification(self.old_report_version, self.new_report_version)
-
+        # ASSERT: Verify that no ReportVerification has been created for the new report version.
         new_verification = ReportVerification.objects.filter(report_version=self.new_report_version).first()
         self.assertIsNone(new_verification)
 
@@ -657,6 +398,7 @@ class ReportSupplementaryVersionServiceTests(TestCase):
         Test that clone_report_version_attachments clones all ReportAttachment instances
         from the old report version to the new report version.
         """
+        # PRE-ACT: Create two ReportAttachment instances for the old report version.
         with pgtrigger.ignore('reporting.ReportAttachment:immutable_report_version'):
             make_recipe(
                 'reporting.tests.utils.report_attachment',
@@ -674,19 +416,19 @@ class ReportSupplementaryVersionServiceTests(TestCase):
                 attachment_name="Attachment 2",
             )
 
+        # ACT: Clone the ReportAttachment instances.
         with patch("django.core.files.storage.default_storage.duplicate_file") as mock_duplicate:
             mock_duplicate.return_value = "test_file"
 
-            ReportSupplementaryCloning.clone_report_version_attachments(
-                self.old_report_version, self.new_report_version
-            )
-
+            clone_report_version_attachments(self.old_report_version, self.new_report_version)
             current_year = timezone.now().year
+            # we add random suffixes to filenames, so we check patterns instead of exact matches
             assert len(mock_duplicate.mock_calls) == 2
             call_args = [str(call_obj) for call_obj in mock_duplicate.mock_calls]
             assert any(f"report_attachments/{current_year}/file1" in arg and ".pdf" in arg for arg in call_args)
             assert any(f"report_attachments/{current_year}/file2" in arg and ".doc" in arg for arg in call_args)
 
+        # ASSERT: Verify that two ReportAttachment instances have been cloned.
         new_attachments = ReportAttachment.objects.filter(report_version=self.new_report_version)
         self.assertEqual(new_attachments.count(), 2)
         for old_attachment in ReportAttachment.objects.filter(report_version=self.old_report_version):
@@ -699,18 +441,20 @@ class ReportSupplementaryVersionServiceTests(TestCase):
     def test_clone_facility_with_activities(self):
         """
         Test that clone_report_version_facilities clones a FacilityReport (with associated ReportActivity)
-        from the old report version to the new report version.
+        from the old report version to the new report version using recipes.
         """
+        # PRE-ACT: Assert that there is exactly one FacilityReport for the old report version.
         initial_count = FacilityReport.objects.filter(report_version=self.old_report_version).count()
         self.assertEqual(initial_count, 1, "There should be one FacilityReport initially.")
 
-        ReportSupplementaryCloning.clone_report_version_facilities(self.old_report_version, self.new_report_version)
+        # ACT: Clone the FacilityReport (and its associated ReportActivity) to the new report version.
+        clone_report_version_facilities(self.old_report_version, self.new_report_version)
 
-        new_count = FacilityReport.objects.filter(
-            report_version__in=[self.old_report_version, self.new_report_version]
-        ).count()
+        # ASSERT: Verify that a new FacilityReport has been created (record count increases from 1 to 2).
+        new_count = FacilityReport.objects.count()
         self.assertEqual(new_count, 2, "A new FacilityReport should be created after cloning.")
 
+        # ASSERT: Retrieve the cloned FacilityReport and verify that its scalar fields match the original.
         new_facility_report = FacilityReport.objects.filter(report_version=self.new_report_version).first()
         self.assertIsNotNone(new_facility_report, "A cloned FacilityReport should exist for the new report version.")
         self.assertEqual(new_facility_report.facility, self.old_facility_report.facility)
@@ -723,6 +467,7 @@ class ReportSupplementaryVersionServiceTests(TestCase):
             new_facility_report.activities.all(), self.old_facility_report.activities.all(), ordered=False
         )
 
+        # ASSERT: Verify that the ReportActivity ReportRawActivityData is cloned.
         cloned_report_activity = ReportActivity.objects.filter(facility_report=new_facility_report).first()
         self.assertIsNotNone(cloned_report_activity, "Cloned ReportActivity should exist for the new FacilityReport.")
         self.assertEqual(
@@ -740,6 +485,7 @@ class ReportSupplementaryVersionServiceTests(TestCase):
         )
 
     def test_reapply_emission_categories(self):
+        # Arrange
         fake_version = MagicMock(spec=ReportVersion)
         fake_methodology = MagicMock(spec=ReportMethodology)
         fake_emission = MagicMock(spec=ReportEmission)
@@ -747,15 +493,17 @@ class ReportSupplementaryVersionServiceTests(TestCase):
         fake_emission.report_fuel = MagicMock()
         fake_emission.report_methodology = fake_methodology
 
+        # Patch the queryset
         with (
             patch("reporting.models.ReportEmission.objects.filter") as mock_filter,
             patch(
-                "reporting.service.report_supplementary_version_service.report_supplementary_version_service.EmissionCategoryMappingService.apply_emission_categories"
+                "reporting.service.report_supplementary_version_service.report_supplementary_cloning.EmissionCategoryMappingService.apply_emission_categories"
             ) as mock_apply,
             patch(
-                "reporting.service.report_supplementary_version_service.report_supplementary_version_service.model_to_dict"
+                "reporting.service.report_supplementary_version_service.report_supplementary_cloning.model_to_dict"
             ) as mock_model_to_dict,
         ):
+
             mock_filter.return_value.select_related.return_value = [fake_emission]
             mock_model_to_dict.return_value = {
                 'report_source_type': fake_emission.report_source_type,
@@ -763,8 +511,10 @@ class ReportSupplementaryVersionServiceTests(TestCase):
                 'report_methodology': fake_methodology,
             }
 
-            ReportSupplementaryVersionService.reapply_emission_categories(fake_version)
+            # Act
+            reapply_emission_categories(fake_version)
 
+            # Assert
             mock_filter.assert_called_once_with(report_version=fake_version)
             mock_apply.assert_called_once_with(
                 report_source_type=fake_emission.report_source_type,
@@ -776,3 +526,35 @@ class ReportSupplementaryVersionServiceTests(TestCase):
                     'report_methodology': fake_methodology,
                 },
             )
+
+    def test_clone_all(self):
+        """
+        Test that clone_all delegates to each individual cloning function
+        with the correct source and target arguments, and calls reapply_emission_categories
+        on the target version.
+        """
+        cloning = "reporting.service.report_supplementary_version_service.report_supplementary_cloning"
+        with (
+            patch(f"{cloning}.clone_report_version_operation") as mock_op,
+            patch(f"{cloning}.clone_report_version_representatives") as mock_reps,
+            patch(f"{cloning}.clone_report_version_person_responsible") as mock_person,
+            patch(f"{cloning}.clone_electricity_import_data") as mock_electricity,
+            patch(f"{cloning}.clone_report_version_additional_data") as mock_additional,
+            patch(f"{cloning}.clone_report_version_new_entrant_data") as mock_new_entrant,
+            patch(f"{cloning}.clone_report_version_verification") as mock_verification,
+            patch(f"{cloning}.clone_report_version_attachments") as mock_attachments,
+            patch(f"{cloning}.clone_report_version_facilities") as mock_facilities,
+            patch(f"{cloning}.reapply_emission_categories") as mock_reapply,
+        ):
+            clone_all(self.old_report_version, self.new_report_version)
+
+            mock_op.assert_called_once_with(self.old_report_version, self.new_report_version)
+            mock_reps.assert_called_once_with(self.old_report_version, self.new_report_version)
+            mock_person.assert_called_once_with(self.old_report_version, self.new_report_version)
+            mock_electricity.assert_called_once_with(self.old_report_version, self.new_report_version)
+            mock_additional.assert_called_once_with(self.old_report_version, self.new_report_version)
+            mock_new_entrant.assert_called_once_with(self.old_report_version, self.new_report_version)
+            mock_verification.assert_called_once_with(self.old_report_version, self.new_report_version)
+            mock_attachments.assert_called_once_with(self.old_report_version, self.new_report_version)
+            mock_facilities.assert_called_once_with(self.old_report_version, self.new_report_version)
+            mock_reapply.assert_called_once_with(self.new_report_version)
